+++ Server/recycleApp/recycleApp/src/main/java/com/recycle/recycleapp/dtos/WasteDTO.java
@@ -6,7 +6,6 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-@Entity
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
