+++ Server/recycleApp/recycleApp/src/main/java/com/recycle/recycleapp/dtos/entryDtos/EntryDto.java
@@ -1,4 +1,4 @@
-package com.recycle.recycleapp.dtos.responseDtos.entryDtos;
+package com.recycle.recycleapp.dtos.entryDtos;
 
 public class EntryDto {
 }
