+++ Server/recycleApp/recycleApp/src/main/java/com/recycle/recycleapp/entities/Waste.java
@@ -1,4 +1,4 @@
-package com.recycle.recycleapp.entity;
+package com.recycle.recycleapp.entities;
 
 
 import com.recycle.recycleapp.enumerations.typeEnum;
