+++ Server/recycleApp/recycleApp/src/main/java/com/recycle/recycleapp/repositories/WasteRepo.java
@@ -1,8 +1,9 @@
 package com.recycle.recycleapp.repositories;
 
+import com.recycle.recycleapp.entities.Waste;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface WasteRepo extends JpaRepository<com.recycle.recycleapp.entity.Waste,Long> {
+public interface WasteRepo extends JpaRepository<Waste,Long> {
 }
